added a page for admin, moved the inventory there, created a page for login add inoput fields
